Introduce new API for navigating around sunspec devices.
The API is configured by the SMDX models and provides a uniform
mechanism for navigating around a strongly typed representation
of a Sunspec device.

It is intended that different that same API can be used irrespective
of whether the actual device is a simulated (in-memory) device, a Modbus
connected device or an XML document containing a copy of device data.

The API knows about scaling factors and helps ensures that scaling factors
and values are read from the underlying device in the same call, relieving
the programmer of some of this complexity.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>