Permit a Device to have multiple models with the same ID.
The SunSpec Model Data Interchange specification (12021) permits
a single device to have multiple instances of a model.

The API previously had this method:
type Device interface {
    Model(id) (Model, error)
}

which contained an implicit assumption that there was only one
model of each type per device.

So, now we remove the Model(id) method and instead require the programmer
to assert there is exactly one model of the specified type or cope
with multiple models by using the Do() function instead.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>