impl: extend set of implied points
The previous implementation led to situations where Block.Read("A") followed by
Block.Read("B") would cause point A to be (surprisingly) invalidated after
the second Read occurred in the case where "A" and "B" are related to the same scale factor
point "C".

With this change, the second read for "B" will imply the read of "C"
(previous behaviour) but now also "A". This ensures that points that were
previously valid remain valid although perhaps with new values.

Of course, it would always be more efficient to only do block.Read("A", "B")
(which would also read "C") but to know this the programmer needs to keep
track of the transitive dependency between "A" and "B" which may not always
be convenient to do.