modbus: Introduce support for modbus physical device implementations.
The library uses the Client interface of github.com/goburrow/modbus
library for the RTU &/or TCP comms. A simulator is provided which
uses a memory slab to simulate the persistence of a real Modbus device.

The key considerations in the implementation of this driver are to
optimise both the read and write schedules to reduce the number of
calls to the modbus interface which, in the case of writes, preserving
the ordering of the writes. Like the memory driver the devices and models,
present in the address space on connection need to be enumerated and the
number of blocks calculated from the size of the models.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>