xml: implement support for Open, Write and Read.
There is now a passing test that round trips from a byte slab into an XML-backed model
(and hence and XML element) and back into an XML based model from the XML element.

ScaleFactor support needs to be tweaked to properly handle
inconsistencies between points and their scale factors in the XML
document.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>