xml: add special case handling for common model (model 1)
The common model, model 1, needs to special case handling by the XML
driver since several of its points are duplicated in the device
element and implementations are allowed to omit it under certain
circumstances.

We attempt to hide this detail from the consumers of the API by
creating an empty (but detached) model 1 element if none has been provided
(during OpenDevice), transparently mirroring updates to
the points Md, Mn, SN into attributes of the the device element
and transparently attaching the detached model 1 element
into the parent device on the first attempt to write some other point
into model 1.