Abstract out AddressLayout and AddressSpace function.
The purpose of these changes is to support non-SunSpec address
spaces with the same API. The reason is that in any practical SunSpec
work, it will be necessary to access parts of the address space
that is not layed out with SunSpec conventions - usually legacy, vendor
specfic address maps that have not been standardised by SunSpec and
have not been developed in accordance with SunSpec design guidelines
(e.g. they are not self-describing).

Now, we could create a whole different API for this purpose, but this API
already does a vast majority of the heavy lifting with regard to meta-data
management and data marshalling, so it seems sensible to enable re-use of this work.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>